Move everything around and setup the CMakeLists.txt.
This is mostly to get it into a format I'm used to dealing with and to make it match everything else in my project.
I assume that the makefile worked/works perfectly fine.

Regenerated the protobuffer files because I was getting an error about the files being created with an old version of protobuf.